# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Amir Numerique - Digital Printing Agency API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin gets all, user gets own)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /portfolio:
    get:
      operationId: listPortfolio
      tags: [portfolio]
      summary: List portfolio items
      parameters:
        - name: featured
          in: query
          schema:
            type: boolean
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of portfolio items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioItem"
    post:
      operationId: createPortfolioItem
      tags: [portfolio]
      summary: Create portfolio item (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortfolioItemBody"
      responses:
        "201":
          description: Portfolio item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioItem"
  /portfolio/{id}:
    get:
      operationId: getPortfolioItem
      tags: [portfolio]
      summary: Get portfolio item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Portfolio item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioItem"
    put:
      operationId: updatePortfolioItem
      tags: [portfolio]
      summary: Update portfolio item (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePortfolioItemBody"
      responses:
        "200":
          description: Portfolio item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioItem"
    delete:
      operationId: deletePortfolioItem
      tags: [portfolio]
      summary: Delete portfolio item (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    RegisterBody:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 8
        preferredLanguage:
          type: string
          enum: [fr, ar]
          default: fr
      required:
        - fullName
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum: [visitor, client, subcontractor, admin]
        preferredLanguage:
          type: string
          enum: [fr, ar]
        accountStatus:
          type: string
          enum: [active, suspended]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - email
        - role
        - preferredLanguage
        - accountStatus
        - createdAt
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
    Service:
      type: object
      properties:
        id:
          type: integer
        nameFr:
          type: string
        nameAr:
          type: string
        slug:
          type: string
        descriptionFr:
          type: string
          nullable: true
        descriptionAr:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        publicPricePerM2:
          type: number
        clientPricePerM2:
          type: number
        subcontractorPricePerM2:
          type: number
        requiresFileUpload:
          type: boolean
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nameFr
        - nameAr
        - slug
        - publicPricePerM2
        - clientPricePerM2
        - subcontractorPricePerM2
        - requiresFileUpload
        - active
        - createdAt
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        userId:
          type: integer
        serviceId:
          type: integer
        widthInput:
          type: number
        heightInput:
          type: number
        unitInput:
          type: string
          enum: [cm, m]
        widthM:
          type: number
        heightM:
          type: number
        areaM2:
          type: number
        unitPricePerM2:
          type: number
        displayedPrice:
          type: number
        finalPrice:
          type: number
        note:
          type: string
          nullable: true
        fileUrl:
          type: string
          nullable: true
        paymentMethod:
          type: string
        paymentStatus:
          type: string
          enum: [pending_on_delivery, paid, cancelled]
        orderStatus:
          type: string
          enum: [pending, confirmed, in_progress, printing, ready, delivered, cancelled]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
          nullable: true
        service:
          $ref: "#/components/schemas/Service"
          nullable: true
      required:
        - id
        - orderNumber
        - userId
        - serviceId
        - widthInput
        - heightInput
        - unitInput
        - widthM
        - heightM
        - areaM2
        - unitPricePerM2
        - displayedPrice
        - finalPrice
        - paymentMethod
        - paymentStatus
        - orderStatus
        - createdAt
        - updatedAt
    CreateOrderBody:
      type: object
      properties:
        serviceId:
          type: integer
        widthInput:
          type: number
        heightInput:
          type: number
        unitInput:
          type: string
          enum: [cm, m]
        note:
          type: string
        fileUrl:
          type: string
      required:
        - serviceId
        - widthInput
        - heightInput
        - unitInput
    UpdateOrderStatusBody:
      type: object
      properties:
        orderStatus:
          type: string
          enum: [pending, confirmed, in_progress, printing, ready, delivered, cancelled]
        paymentStatus:
          type: string
          enum: [pending_on_delivery, paid, cancelled]
      required:
        - orderStatus
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - orders
        - total
        - page
        - limit
...
    PortfolioItem:
      type: object
      properties:
        id:
          type: integer
        titleFr:
          type: string
        titleAr:
          type: string
        descriptionFr:
          type: string
          nullable: true
        descriptionAr:
          type: string
          nullable: true
        imageUrl:
          type: string
        category:
          type: string
          nullable: true
        isFeatured:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - titleFr
        - titleAr
        - imageUrl
        - isFeatured
        - createdAt
    CreatePortfolioItemBody:
      type: object
      properties:
        titleFr:
          type: string
        titleAr:
          type: string
        descriptionFr:
          type: string
        descriptionAr:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        isFeatured:
          type: boolean
          default: false
      required:
        - titleFr
        - titleAr
        - imageUrl
    UpdatePortfolioItemBody:
      type: object
      properties:
        titleFr:
          type: string
        titleAr:
          type: string
        descriptionFr:
          type: string
        descriptionAr:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        isFeatured:
          type: boolean
...