# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Amir Numerique - Digital Printing Agency API
 * OpenAPI spec version: 0.1.0
 */
import type { UserAccountStatus } from "./userAccountStatus";
import type { UserPreferredLanguage } from "./userPreferredLanguage";
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  fullName: string;
  email: string;
  phone?: string | null;
  role: UserRole;
  preferredLanguage: UserPreferredLanguage;
  accountStatus: UserAccountStatus;
  createdAt: Date;
}